Remove projectPatchResolver from project.update hooks
projectPatchResolver has a converter for the data. project.update
currently has custom parameters, where the data is the first parameter,
not the second, which is usually null. This was causing errors when trying
to get `settings` off of it, and even returning null caused problems, as
the converter expects an object of some sort.