Fixed issue with pushing project to empty GitHub repo
Backend code was trying to delete an empty repo and then crete it with auto_init
set to true. This needs the `admin` GH scope, though, which is not requested.
Changed this to creating a dummy readme file instead.

Fixed bug in edit ProjectsPage that was not refreshing the installed projects on
project patch. This was causing projects that were patched to have a GitHub repo
link to not have their options updated to reflect that.