Fixed infinite recursion in avatar & user resolvers.
PR 9073 introduced an infinite recursion bug. If a user had a custom avatar
that they had uploaded, the user resolver would fetch their avatar, which
would now fetch the user that owned that avatar, which would fetch the
avatar, ad infinitum. Added some query params to skip the population
of the user or avatar if instructed to do so by internal resolver fetches.