Fixed server-side meshopt_decoder Worker call.
Backend running of meshopt_decoder was failing because of lack
of Worker in node. Now polyfilling it with web-worker like we do with
other uses of Worker. Had to change meshopt_decoder Blob URL to Data URL
since web-worker does not appear to work with Blob URLs despite the
documentation saying it does.

Added senderId to 'user left instance' message to avoid a bug. Message
onCRUD expects a senderId if sender is not present. Added a check for
senderId in onCRUD just in case it isn't.