Updated S3 storage provider to make most project files private
When adding files to S3, any project files not in a top-level
`public` folder will have a private ACL so as not to expose
code. Anything within the top-level `public` folder will remain
as public-read, no matter how deep, but `public` folders in other
project folders will not be made public.