Updated file-browser to update project files locally
In local dev, when project files were created/moved/removed,
those operations only occurred in the storage provider, not
in projects/projects/<project>. For those doing local work
on projects and wanting to push from their local file system,
this meant that no changes were being recorded.

file-browser operations now, if isDev is true and acting on
project files, make the equivalent operation on the local
file system as well.

Fixed bugs in the creation of new scenes. One, the operation
was trying to copy a default.scene.json from the project being
operated on, not the default project. Two, the path of the
output files was missing a slash.