Force db refresh on builder
Using a different ENV_VAR than FORCE_DB_REFRESH since the builder configmap
starts with the builder's extraEnv, then overwrites it with api/client/instanceserver
values. If builder's FORCE_DB_REFRESH is true but the api's is false, then it'll end
up being false on the configmap. If we set the api's to true, then the api servers
will do their own refresh, which we don't want.