Tweaked FileBrowserContentPanel to allow better setting of nesting directory
In the studio, we want to let people go to other projects through the file browser,
while in other situations we do not want to let them go above the level they start
at, e.g. projects/<projectName> or /recordings/<recordingID>. Added a new prop to
the panel component to optionally set a nesting directory. Tweaked the breadcrumbs
logic to not show anything higher than the nesting directory.