Fixed bug with user disconnect on world instanceserver.
If a user disconnected because they had signed in from being
a guest, and the guest user was then destroyed because the
logged-in user already existed, the instanceserver's attempt
to create a 'user left' message was throwing an error because
looking up that user by ID returned nothing. Now catching that
and creating a generic user left message since their name
is now impossible to find, with a null senderId because messages
can't have a senderId of a non-existent user.