Added hasLocalChanges to project to prevent data loss
Data loss in projects was occurring when projects were set to
auto-update, changes were made to that project on a deployment,
and those changes were not pushed to GitHub.

To avoid this, now when anything in a project is saved, it will
set hasLocalChanges to true on that project. When a project
auto-updater runs, it will not update the project if hasLocalChanges
is true. Pushing the project to GitHub will clear hasLocalChanges,
as will manually updating the project.