Fixed inefficiency in resolving project's permissions
All project methods were trying to resolve a project's permissions.
This would in turn require a couple additional database lookups.
When many projects were trying to be fetched at once, this would
make the overall project.find take several seconds.

The only place that project permissions are currently used is on the
main studio page, and they're not displayed or edited unless the user
has opened the context modal for a specific project.

Adding a project's permissions in the resolver has been made an opt-in
query param which gets moved to context.params from context.params.query.
Otherwise, it just returns an empty array, speeding the process up
quite a bit.

The studio main page has been altered so that when a project's context
is loaded, it fetches just that project's permissions and manually adds
them onto the project.