Fixed bugs with builder commit fetching (#10156)
* Fixed bugs with builder commit fetching

builder-info.class.ts was using the wrong value from privateECRRegexExec.

findBuilderTags in project-helper.ts was not using the new fromIni
method to get credentials for AWS. Updated it to do so.

Made S3 and EKS SDKs save their credentials files to separate files,
so that they don't collide with each other.

* Don't write EKS credentials file if it exists.