Made client's server.js not bound to a specific host (#10223)
The client's app.listen() call was trying to listen to the host VITE_APP_HOST. In a minikube
environment, this would error out since that address had no DNS routing internal to minikube,
and the client pod would fail, as the listen() call attempts to ping the host while starting,
and throws an error if it cannot do so.

Removing the host from app.listen just has it listen on 0.0.0.0, which will always succeed.