Fixed bug in scope creation when scope exists (#10260)
When creating a scope for a user who already had that scope, existing logic
was throwing errors in feathers about 'Cannot redefine propery: Symbol'. It's
unclear why this was being thrown since it dealt with feathers internals, but
simplifying the checkExistingScopes logic to just remove existing scopes instead
of trying to track which ones already existed, like is already done when patching
a user with new scopes, solved the problem.