Updated Cloudfront function to pass most things as routes (#10234)
Project-defined routes were not being handled properly by the Cloudfront function;
the function was not properly handling subroutes of the registered route, e.g.
/newroute/subroute would throw a 403 because the CF function didn't know how to
handle anything other than /newroute.

Rather than make the function conditionally handle subroutes, it was determined that
it would be easier to instead have /projects and /recordings paths be passed through
as-is, and have everything else that wasn't a static client file be passed as if it
were a route, i.e. get directed to client/index.html and let the router there determine
what to do with it (or nothing, if it's not a registered route).

Co-authored-by: Hanzla Mateen <hanzlamateen@live.com>