Support project update via GH app JWT
Implements functionality for IR-2465

If project.update is called with a GitHub App JWT in the body (passed
as 'token'), this will trigger alternative logic in the authenticate hook.
It will check that the token comes from the same GH App that's registered
with the deployment and that the token is valid. This will bypass user
authentication, and put the token on context.params.appJWT.

updateProject will use this token to generate an installation access token
for the repo the project is in (throwing an error if the app is not
installed such that it can access that repo), then use that token to do
the git clone instead of a user OAuth token.

Updated GitHub admin authentication-settings to allow for entry of AppID.
Also updated OAuth authentication-setting schema to account for appId.

Fixed some minor bugs with project name usage when calling project.update
from the front-end.