Removed font.css, replaced with meta tag imports in components (#10328)
Import of Google fonts in font.css was breaking first line of later .css file
imports because of lack of semicolon at the end. But we should be importing fonts
in components via <link>s hoisted into the <head> via <MetaTags>. This fixes IR-2428.

When the user menu is rendered in a modal, e.g. at /studio,its inclusion
in an MUI modal was adding some undesired styling, specifically a white
background. Made the menu stylings all !important to override that behavior.
This fixes IR-2452.