Added cache busting to studio asset requests
Users were encounter issues in the studio where saving a scene, then immediately
refreshing it, could result in the cached old version of the scene (or assets)
being retrieved because the cache invalidation had not been resolved yet.

Made a new function to add a unique hash onto the query params of asset fetches
if in the editor. This will bust the cache every time so that this stale loading
does not occur. Currently the hash is epoch time in seconds; in the future it
should be the hash of the file if that information is easily available to the
client (retrieving it right now would be cumbersome and slow).