Remove react-is hackfix install
Recent updates to typescript and eslint versions made the hackfix install
of react-is in building the client error out, and seems to no longer be
necessary to get it to install properly.

Moved @emotion/react and @emotion/styled to main dependencies from
devDependencies since the client build was throwing errors about them
being missing.