Upgraded WebRTC options (#10555)
Added webRTCSettings column to instance-server-settings table.
This allows for configuration of custom iceServers definition, as
well as returning raw IP addresses for private instanceserver addresses
instead of private DNS records, and using the TURN REST API to generate
temporary credentials for users to access a TURN server.

This allows for using a TURN server to relay WebRTC connections, so that
instanceservers can be placed behind private subnets and still be reachable.
This will also work inside a VPN as a STUN server, and can still make direct
connections.

iceServers definition is now always provided to the client from the instanceserver
over the websocket connection when making a transport. This allows the
instanceserver to generate credentials for the TURN REST API from the shared
secret without the client knowing anything about the workings of that process.
It just gets told where to connect to for STUN/TURN, and any credentials if
that's relevant.

Resolves IR-1445 and all associated issues.