Hardened instanceserver initialization
Instanceserver initialization was making new instance records
if it couldn't find one that matched the IP address and
location/channel ID provided by the connecting user. This could
be abused by a malicious actor. Now, the instanceserver just
checks if there is a record for itself, and does nothing if it
can't find one.

Moved the initial authorizeUserToJoinServer call earlier so that
nothing else happens if the user shouldn't be on the server in the
first place. loadEngine is not called if the server isn't initialized
for any reason. User doesn't join the instanceserver if not authorized
to do so. Cleaned up some duplicate logic.

Resolves IR-2671