Disabling screenshare if it's not supported.
Mobile browsers do not support navigator.mediaDevices.getDisplayMedia,
which is how one retrieves a display for screensharing over WebRTC.
getDisplayMedia is undefined on all mobile browsers (or should be - older
versions of most browsers it was exposed, but returned a NotAllowedError).
Now hiding the screenshare button if isMobile is true or getDisplayMedia
is undefined.

Resolves IR-3136