Fixed issues with user media.
User video for producers that existed before a client joined
an instance was not getting hidden on producer pause because
there was no local record of that producer to mark as paused.
Added a producerPaused field to consumers so that this status
can be held and <video/audio>ProducerPaused can be updated
off of that. Initial paused state of consumers is now always
false.

Removed NetworkProducerConsumerPause because it was causing
problems. If a producer unpauses, it should not force-resume
its consumers. Users may have paused their consumer locally,
and this was overriding their pause.

Rendering FlipVideo button only if there is more than one
video device.

Resolves IR-3222