Added refreshing of GitHub user OAuth tokens. (#10716)
GitHub apps by default generate short-lived (8 hours) OAuth tokens,
and require refreshing of them via a refresh token that is generated
as part of the OAuth flow. The assumption had been that the tokens
were long-lived, leading to people needing to sign out and in often.

Added a column to identity-provider to hold a user's refresh token.
Updated GitHub strategy to pass this refresh token and save it on
the identity provider.

Updated GitHub call flow to check if the token to be used is valid.
If not, then it is refreshed using the user's refresh token.
Refresh tokens are invalidated when used once, but they return a
new refresh token that is saved on the identity provider.

Had to pass `app` to most github-helper.ts calls in order to support
the refresh process, as it requires the GH app public and private keys,
which have to be retrieved from the database.

Fixed some visual bugs on the admin add/edit project modal.

Resolves IR-3325

Co-authored-by: Hanzla Mateen <hanzlamateen@live.com>