Fixes for instanceserver connection bugs
A rare situation was cropping up where an instance record was
getting removed from the database after an instance provision.
This would cause initializeInstance to return false, and loadEngine
to not run, but instanceStarted would remain true. That instance
wouldn't start, but the server would still remain in the pool of
Ready servers. Later clients assigned to that pod would then just
endlessly wait for InstanceServerState.ready to be true, which
would never happen. setting instanceStarted back to false if
initializeInstance returns false should fix this.

Increased the timeout on assigned but not-connected instances to
60 seconds from 30. This may be what was causing those instance
records to be removed, if a client on a slow network was getting
the provision but then taking more than 30 seconds to make the
websocket connection.