Fixed bugs in instanceserver connections (#10887)
* Fixed bugs in instanceserver connections

There were still some bugs surrounding a client connecting to an
instanceserver if the instance record for that server no longer
existed. updateInstance needed to return false in that situation.

This still led to a situation where the client would be stuck trying
to authenticate forever. Added a step prior to authentication where
the client listens for a message with { instanceReady: true }, which
is sent by setupSocketFunctions. If the server isn't ready within
10 seconds, then it returns { instanceReady: false }, and the client
will do a new instanceserver provision and hopefully connect properly
this time.

* make eslint happy

---------

Co-authored-by: Josh Field <joshfield999@gmail.com>