Fixed bugs in identity-provider service (#10917)
Users were able to create an identity-provider on another user.
This would return a valid JWT for that other user that could
authenticate requests as them. Enhanced identity-provider hooks
for creation so that users can only create identity-providers
for themselves, and unauthenticated requests cannot create
identity-providers on existing users.

Also prevented the creation of guest identity-providers on
existing users, because there's no reason for it and it seems
like something that could introduce bugs.

Added tests to ensure that one cannot create an identity-provider
on another user. Updated a number of tests to work around some
issues this causes - oauth providers are seen as 'guest' providers,
so tests that were creating github sign-ins were now failing.

Some identity-provider tests were not actually running to completion.
assert.rejects needs to take in an async function whose contents are
await'ed, and the assert.rejects itself needs to be await'ed, or else
the test suite will mistakenly assume it succeeded when it's still
running.