Made github-repo-access-refresh runnable as a job
Logging in with GitHub was taking notably longer than other
SSO methods because it was performing a refresh of the user's
Github repo access. Made a new service to call this function
that can be run as a job, and running it as a job on Kubernetes.

createExecutorJob now has an optional param to not wait for the
job to finish before resolving for this and similar situations
where you don't care to check if the job is finished. Default
is to waitForFinish to be true.