Moved ECR pruning to after all build services have succeeded.
Prior behavior was to prune images for a service after that
service had built successfully. This still had holes in the
logic of which ones to remove, and sometimes several builds
that had e.g. a failing client build would prune still-in-use
api images, leading to the cluster being in a state where it
had to run an older version of the code but the images with
that code were gone.

Now, prune_ecr_images is only run if every build has succeeded,
which should prevent any pruning of images that need to still
be used.

Resolves IR-3455