Updated schema of webRTCSettings.iceServers
Credentials and whether to use fixed or time-limited credentials
are now stored on each iceServer record. This makes it not JSON-
standard with iceServers, but it is converted to the right schema
when populating iceServers on the instanceserver.

Standardized on containerClassName for UI elements (some had been
using containerClassname, others containerClassName)

Updated Label element to accpt hmtlFor, and to apply that, so that
labels on inputs can be clicked and change the input