Fixed a bug in creating new guest users (#93)
If a now-invalid JWT was signing a request to make a new guest user with a userId of
empty string, it was throwing an error when checking if that userID matched the userId
of the JWT. This would cause the client to hang indefinitely until a refresh. Now
specifically excluding userId of empty string from that check, as that will always
create a new guest user, which is not a security hole.