Fixed a bug with consumer state change. (#209)
In intermittent situations, usually with >=4 users connected to a
media instancserver, one user leaving would put every producer in
an un-consumable state. The main error showing up was that some
consumers' producerID would be undefined, and trying to get the
.value of it was throwing errors, and possibly triggering all
of the other errors. Made that value read conditional and returning
if consumer.producerID is null/undefined; hopefully this solves it.

Resolves IR-4431