Updated authentication pattern
Initial auth request is done via a REST request. This is so that the JWT
can be added onto the headers of the Primus socket request.

ClientAPI.createAPI now called from AuthService, as we don't want to create
the websocket connection on page load, rather after we've done the REST request.

Using a modified version of primus-client that will set the token as the
sec-websocket-protocol header, which is the only custom header that can
be set on a websocket request by a browser.