doLoginAuto now calls generate-token to get short-lived JWT for WS
generate-token can now put a policy and custom expiresIn on the token
The tokens generated by login/identity-provider.create/etc. will be
the normal 30-day tokens for verifying a user, but they won't have a
Tyk policy on them, and will just be used to generate the short-lived
tokens, which will be used to sign websocket requests so that Tyk can
do its rate-limiting.