Updated error message thrown by missing OAuth state
Very rarely, the OAuth flow will end up in a state where
there's no profile information. It then tries to read a
value off of this null value and throws an error.

Attempts to diagnose how the OAuth flow ended up in this
state were unsuccessful, as it almost never happens, and
after happening once, never seems to happen again on a
deployment. The solution is to throw an error message that
prompts the user to try logging in again, as that will
almost assuredly work.

Resolves IR-4857