Fixed bug in static-resource patching
FilePropertiesModal was patching static-resources with the project
currently loaded in the studio, which may not be the project of
the file being changed. e.g. if a user is on the project projectA,
but they edit an asset in projectB, it was patching that projectB
asset to belong to projectA. Then, when updating the resources.json
file, it was updating the resources.json of the project of that
asset, which would now be projectA's resources.json. projectB's
resources.json would be left untouched, and projectA's would now
have an entry for an asset that it didn't own.