Fixed renaming of scenes
RenameSceneModal was passing the full path of the scene file,
including the file name, on oldPath and newPath, and also having
the file names be oldName and newName. This was causing the backend
to not match the concatenated file path and not do anything. e.g.
it was trying to move a file named
projects/thisorg/thisproject/public/scenes/newscene.gltf/newscene.gltf to
projects/thisorg/thisproject/public/scenes/newscene2.gltf/newscene2.gltf

Resolve IR-6255