Increased the S3 request handler timeout and maxSockets
The client build process was starting to frequently get clogged when
pushing files to S3, which was leading to index.html not being
pushed, but its child index.js files were getting pruned.

AWS documentation suggested increasing number of sockets in these
situations. Added some logging so if it continues to happen, it's
easier to detect why this isn't being flagged as an error by
record-build-error.