Fixed email matching on new OAuth storage providers
There were some cases, specifically with GitHub, where the profile
of an OAuth provider did not have an email address, but we were
getting it through other means. This was leading to a lack of
matching with existing identity providers.

Now checking using the email of the new identity provider instead
of the email on the profile returned from the OAuth provider.