Clearing guest identity providers on successful login
Certain bugs in OAuth login flow were not removing guest identity providers.
This caused issues in multitenancy project, where the existence of a guest
IP led the login pages to think the user was a guest, even though they had
non-guest identity providers. They would always be redirected back to the
welcome page by mistake instead of being sent to the rest of the signup flow
or to the dashboard.

Resolves IR-6436