Fixed bug with projects deleting similarly-named projects
If one project matched the start of the name of another project, such as
irpro-ecomm and irpro-ecommerce-tools, adding/updating/removing the former
would match when calling deleteProjectFilesInStorageProvider, and erroneously
remove the latter's files as well. Added a slash to the end of the
getFileKeysRecursive call in there to specifically match just that project.