Added retry plugin to octokit
Pushing projects to GitHub was often running into ETIMEDOUT errors when uploading blobs.
This would cause the whole push to error out.
It's unclear why, but it may have been due to rate limiting, since a lot of files were
being uploaded for certain projects.

Enabled the retry plugin for octokit, which will retry failed requests. In limited testing
with projects exhibiting this behavior, this seemed to solve the problem.